Make absolute addressing mode unsigned
It doesn't make sense to have the absolute displacement be signed;
however, this leaves the other modes signed.  The documentation is
a little inconsitent with this variable:

"For the absolute displacement addressing mode, the offset value
ranges from 0 to 2 ^^ 32 -1. The absolute displacement addressing mode
is encoded in the MEMB format."

vs

"For the absolute displacement addressing mode the offset is an
integer (a displacement) ranging from -2 ^^ 31 to 2 ^^ 31 -1. The
absolute displacement addressing mode is encoded in the MEMB format."