Refactor multi-dword access
8-byte, 12-byte, and 16-byte access should be improved.  This
fixes an issue with more than one register being used, but only
having the first register.  Added other macros to make the code
look consistent.

This did lead to issues with lit vs reg, those are now broken up
into two of the same instruction instead of using a constructor.
This might be able to be accomplished with them, but this was
easy fix for now.